build docker images with buildkit in GitLab
Building but not publishing (yet) the docker container images
for mopidy and snapcast-server.  Need to work out some TLS
issues related to using an internal certificate authority
before pushing the images to a registry.

Also need to find a way to eliminate the per-image duplication
in `.gitlab-ci.yml` before it gets (more) overwhelming.